Moved StasticsGraphs to /Debug/Graphs folder.
